requirements.txt
flask~=2.0.2
line-bot-sdk
wheel
grpcio
google-cloud-firestore
flask_cors
line-bot-sdk
google-cloud-storage
pandas~=1.3.4
gunicorn
google-cloud-logging
tensorflow-cpu
Pillow~=8.4.0
flask_ngrok
Pillow
google-cloud-bigquery
google-cloud-bigquery-storage
BigQuery-Python
pyarrow~=6.0.1
google-cloud-speech
numpy~=1.21.4
requests~=2.26.0
jieba~=0.42.1
tensorflow~=2.7.0